In the case where:
1) ProbeManager dies with sigkill or sigsegv
2) daemons are reclaimed but their startup shells are dead
3) ProbeManager dies AGAIN

there will be incorrect pid files for the running daemons but the manager
won't be able to find them.  The end result being the inability to start
the services that are single instance (like cassandra).  Added a
way to kill off these processes and a way to find them better.